fastapi_app/requirements.txt
sqlalchemy
psycopg2
fastapi
python-decouple
